♻️ refactor: 친구 탭 컴포넌트 분리
