+++ app/(route)/(main)/my/page.tsx
@@ -2,21 +2,12 @@
 import { useState, useEffect } from 'react'
 import { auth, db } from '@/app/_utils/firebaseConfig'
 import { onAuthStateChanged, User } from 'firebase/auth'
-import {
-  collection,
-  doc,
-  getDoc,
-  getDocs,
-  query,
-  updateDoc,
-  where,
-} from 'firebase/firestore'
+import { doc, getDoc } from 'firebase/firestore'
 import { useRouter } from 'next/navigation'
 import { motion, AnimatePresence } from 'framer-motion'
 import UserTabBar from '@/app/_components/TabBar/UserTabBar'
-import AddFriendModal from '@/app/_components/modal/AddFriendModal'
-import FriendList from '@/app/_components/FriendList/FriendList'
 import Profile from '@/app/_components/Profile/Profile'
+import Friends from '@/app/_components/Friends/Friends'
 
 type UserData = {
   email: string
@@ -39,10 +30,6 @@ export default function MyPage() {
   const [isEditing, setIsEditing] = useState(false)
   const [activeTab, setActiveTab] = useState<'profile' | 'friends'>('profile')
   const [tempUserData, setTempUserData] = useState<UserData>(userData)
-
-  const [showModal, setShowModal] = useState(false)
-  const [friendNickname, setFriendNickname] = useState('')
-  const [friendError, setFriendError] = useState('')
 
   const router = useRouter()
 
@@ -101,56 +88,6 @@ export default function MyPage() {
     return friendsDetails
   }
 
-  // 친구 추가 로직
-  const handleAddFriend = async () => {
-    if (!friendNickname.trim()) {
-      setFriendError('닉네임을 입력하세요.')
-      return
-    }
-
-    try {
-      const usersRef = collection(db, 'users')
-      const q = query(usersRef, where('nickname', '==', friendNickname))
-      const querySnapshot = await getDocs(q)
-
-      if (querySnapshot.empty) {
-        setFriendError('해당 닉네임의 사용자를 찾을 수 없습니다.')
-        return
-      }
-
-      const friendDoc = querySnapshot.docs[0]
-      const friendId = friendDoc.id
-
-      if (user?.uid === friendId) {
-        setFriendError('자신을 친구로 추가할 수 없습니다.')
-        return
-      }
-
-      if (userData.friends?.includes(friendId)) {
-        setFriendError('이미 친구로 등록된 사용자입니다.')
-        return
-      }
-
-      const updatedFriends = [...(userData.friends || []), friendId]
-      const userDocRef = doc(db, 'users', user?.uid || '')
-      await updateDoc(userDocRef, { friends: updatedFriends })
-
-      setUserData((prev) => ({
-        ...prev,
-        friends: updatedFriends,
-      }))
-      alert(
-        `🎉 친구 추가에 성공했습니다! "${friendNickname}"님과 친구가 되었어요!`,
-      )
-      setShowModal(false)
-      setFriendNickname('')
-      setFriendError('')
-    } catch (error) {
-      console.error('친구 추가 중 오류 발생:', error)
-      setFriendError('친구를 추가하는 데 실패했습니다.')
-    }
-  }
-
   return (
     <div className="mx-auto mt-10 h-screen w-full max-w-[1200px] px-4">
       <div className="flex flex-col rounded-3xl bg-white p-6 shadow-md">
@@ -184,29 +121,11 @@ export default function MyPage() {
               exit={{ opacity: 0, x: -20 }}
               transition={{ duration: 0.3 }}
             >
-              <div className="flex items-start justify-between">
-                <h2 className="mb-4 text-2xl font-semibold">친구 목록</h2>
-                <button
-                  onClick={() => setShowModal(true)}
-                  className="rounded-full bg-green-30 px-6 py-3 text-white shadow-lg transition hover:bg-green-40"
-                >
-                  친구 추가
-                </button>
-              </div>
-              <FriendList friends={userData.friendsDetails || []} />
+              <Friends userData={userData} setUserData={setUserData} />
             </motion.div>
           )}
         </AnimatePresence>
       </div>
-      {showModal && (
-        <AddFriendModal
-          onClose={() => setShowModal(false)}
-          onAddFriend={handleAddFriend}
-          friendNickname={friendNickname}
-          setFriendNickname={setFriendNickname}
-          friendError={friendError}
-        />
-      )}
     </div>
   )
 }
